irctc-captcha-solver/requirements.txt
argparse
pillow
easyocr==1.7.1
numpy
flask
pillow
pytesseract